Standardize directories into unified file
